Split server-parents-booking.test.ts into themed files
Splits the 1265-line monolith test/lib/server-parents-booking.test.ts into
11 themed files under test/lib/server-parents-booking/ (each under ~210
lines), drops it from biome.json noExcessiveLinesPerFile override, and
extracts the shared helpers the split surfaced (jscpd still reports 0%).

Pure relocation: every test name and assertion is unchanged. The new
files each get their own describeWithEnv(...), grouped by coherent theme.

The duplication jscpd surfaced when it could finally scan the smaller
files is fixed with shared helpers in #test-utils — no jscpd:ignore:
- enablePublicApi / enablePublicSite (test-utils/settings.ts)
- bookableStartDates(listingId) (test-utils/db-helpers/listings.ts)
- listingDetail / availabilityJson / apiListingRow / bookParentChild /
  expectChildAvailability / makeParentWithDeactivatedChild /
  soldOutParentInGroup / expectNoListingsCta (test-utils/parents.ts)

All 73 split tests pass; precommit is green (typecheck, lint, cpd 0%,
tests, mutation over the branch diff).
